+++ MAIDOS-CodeQC/maidos-codeqc/src/rules/b-redlines.ts
@@ -4,15 +4,15 @@ export type { Redline } from './b-redlines-defs.js';
 export { REDLINES, getRedline, getImplementedRedlines, getUnimplementedRedlines, getRedlineStats } from './b-redlines-defs.js';
 import { stripRustCfgTestBlocks } from './b-redlines-utils.js';
 
-import { R01_CHECKER, R02_CHECKER, R03_CHECKER, R05_CHECKER } from './b-redlines-r01-r05.js';
-import { R07_CHECKER, R08_CHECKER, R09_CHECKER, R10_CHECKER, R12_CHECKER } from './b-redlines-r07-r12.js';
+import { R01_CHECKER, R02_CHECKER, R03_CHECKER, R04_CHECKER, R05_CHECKER } from './b-redlines-r01-r05.js';
+import { R06_CHECKER, R07_CHECKER, R08_CHECKER, R09_CHECKER, R10_CHECKER, R11_CHECKER, R12_CHECKER } from './b-redlines-r07-r12.js';
 import { R13_CHECKER, R14_CHECKER, R15_CHECKER, R16_CHECKER, R17_CHECKER, R18_CHECKER } from './b-redlines-r13-r18.js';
 import { R19_CHECKER, R20_CHECKER, R21_CHECKER, R22_CHECKER, R23_CHECKER, R24_CHECKER } from './b-redlines-r19-r24.js';
 import { R25_CHECKER, R26_CHECKER, R27_CHECKER, R28_CHECKER } from './b-redlines-r25-r28.js';
 
 export const REDLINE_CHECKERS: RuleChecker[] = [
-  R01_CHECKER, R02_CHECKER, R03_CHECKER, R05_CHECKER,
-  R07_CHECKER, R08_CHECKER, R09_CHECKER, R10_CHECKER, R12_CHECKER,
+  R01_CHECKER, R02_CHECKER, R03_CHECKER, R04_CHECKER, R05_CHECKER,
+  R06_CHECKER, R07_CHECKER, R08_CHECKER, R09_CHECKER, R10_CHECKER, R11_CHECKER, R12_CHECKER,
   R13_CHECKER, R14_CHECKER, R15_CHECKER,
   R16_CHECKER, R17_CHECKER, R18_CHECKER,
   R19_CHECKER, R20_CHECKER, R21_CHECKER, R22_CHECKER, R23_CHECKER, R24_CHECKER,